feat: add terraform for s3 & cloudfront
