+++ cafe/views.py
@@ -1,133 +1,3 @@
 from django.shortcuts import render
 
 # Create your views here.
-import logging
-import requests
-
-from celery import shared_task
-from django.utils import timezone
-
-from cafe.models import Cafe
-
-from decouple import config
-import boto3  # Python SDK
-from io import BytesIO
-
-logger = logging.getLogger(__name__)
-
-
-@shared_task(
-    bind=True,
-    max_retries=3,
-    default_retry_delay=60,  # 60 秒後重試
-    autoretry_for=(Exception,),  # 自動重試所有例外
-)
-def download_and_upload_cafe_photo(self, cafe_id):
-    """
-    異步任務：下載咖啡店照片並上傳到 S3
-
-    Args:
-        cafe_id (int): Cafe 的 primary key
-
-    Returns:
-        dict: 執行結果
-    """
-    try:
-        # 1. 取得咖啡店資料
-        try:
-            cafe = Cafe.objects.get(id=cafe_id)
-        except Cafe.DoesNotExist:
-            logger.error(f' Cafe ID {cafe_id} 不存在')
-            return {'status': 'failed', 'reason': 'cafe_not_found'}
-
-        # 2. 檢查是否有photo_reference
-
-        if not cafe.photo_reference:
-            logger.info(f'跳過，name：{cafe.name}，place_id：{cafe.place_id}，沒有 photo_reference')
-            return {'status': 'skipped', 'reason': 'no_photo_reference'}
-
-        # 3. 檢查該咖啡店的照片是否上傳至 S3
-
-        if cafe.photo_s3_url:
-            logger.info(f'跳過，name：{cafe.name}，place_id：{cafe.place_id}，已經上傳照片至s3')
-            return {'status': 'skipped', 'reason': 'already_uploaded'}
-
-        logger.info(f'開始下載照片：{cafe.name} (place_id：{cafe.place_id})')
-
-        # 4. 從 Google 下載照片
-        photo_url = (
-            f'https://maps.googleapis.com/maps/api/place/photo'
-            f'?photo_reference={cafe.photo_reference}'
-            f'&maxwidth=400'
-            f'&key={config("GOOGLE_API_KEY")}'
-        )
-
-        response = requests.get(photo_url, timeout=10, stream=True)
-        response.raise_for_status()
-
-        # 檢查 content-type
-        content_type = response.headers.get('Content-Type', '')
-        if 'image' not in content_type:
-            logger.warning(f'非圖片類型：{content_type}')
-            return {'status': 'failed', 'reason': 'invalid_content_type'}
-
-        # 4. 上傳到 S3
-        s3_client = boto3.client(
-            's3',
-            aws_access_key_id=config('AWS_ACCESS_KEY_ID'),
-            aws_secret_access_key=config('AWS_SECRET_ACCESS_KEY'),
-            region_name=config('AWS_REGION', default='ap-northeast-1')
-        )
-
-        bucket_name = config('S3_BUCKET_NAME')
-        s3_key = f'cafe-photos/{cafe.place_id}.jpg'
-
-        logger.info(f'上傳到 S3：{s3_key}')
-
-        s3_client.upload_fileobj(
-            BytesIO(response.content),
-            bucket_name,
-            s3_key,
-            ExtraArgs={
-                'ContentType': 'image/jpeg',
-                'CacheControl': 'max-age=31536000',  # 1 年快取
-                # Terraform 設定 S3 為拒絕私人連線
-            }
-        )
-
-        # 5. 生成 URL
-        cloudfront_domain = config('CLOUDFRONT_DOMAIN', default='')
-
-        if cloudfront_domain:
-            # 使用 CloudFront（快取）
-            s3_url = f'https://{cloudfront_domain}/{s3_key}'
-        else:
-            # 直接用 S3 URL
-            region = config('AWS_REGION', default='ap-northeast-1')
-            s3_url = f'https://{bucket_name}.s3.{region}.amazonaws.com/{s3_key}'
-
-        # 6. 更新資料庫
-        Cafe.objects.filter(id=cafe.id).update(
-            photo_s3_url=s3_url,
-            photo_updated_at=timezone.now()
-        )
-
-        logger.info(f'成功上傳照片至 S3，Cafe name：{cafe.name} → {s3_url}')
-
-        return {
-            'status': 'success',
-            'cafe_id': cafe.id,
-            'cafe_name': cafe.name,
-            'place_id': cafe.place_id,
-            's3_url': s3_url
-        }
-
-    except requests.RequestException as e:
-        logger.error(f' 下載照片失敗：{e}')
-        # Celery 會自動重試
-        raise
-
-    except Exception as e:
-        logger.error(f'上傳 S3 失敗：{e}', exc_info=True)
-        # Celery 會自動重試
-        raise
