+++ terraform/provider.tf
@@ -1,0 +1,32 @@
+terraform {
+  required_version = ">= 1.0"
+
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = "~> 5.0"
+    }
+  }
+
+  # 遠端 State 儲存
+  backend "s3" {
+    bucket         = "find-coffee-terraform-state"
+    key            = "production/terraform.tfstate"
+    region         = "ap-northeast-1"
+    encrypt        = true
+
+    # 防止多人同時操作，操作時會鎖住
+    dynamodb_table = "terraform-state-lock"
+  }
+}
+
+provider "aws" {
+  region = var.aws_region
+  default_tags {
+    tags = {
+      Project     = var.project_name
+      Environment = var.environment
+      ManagedBy   = "Terraform"
+    }
+  }
+}
